Implement QuoteService for loading data
This commit starts implementing a new QuoteService responsible for logic
implementation of data related to Quotes. For now this service deals
with the actions that will be done at the start of the application - clean
and load data. The load data process is done via RestTemplate which
allows making requests to the Quotes API.